Fixed typo
The crypto all seems to work now.
Not sure what the webworker is for - will return to it once I figure that out.

Amazingly, the only thing left is a regular old npm module.
Why the unexpected behavior? Is Long the issue?
How can I force Long in browserify, if so?